Configure menu for animation
